XXX Hack to allow tor-fw-helper to be built on Windows
For some reason, --with-libminiupnpc-dir doesn't work on Windows, so this
hardcodes /local/lib as the path in which libminiupnpc.a can be found.
Also, libminiupnpc needs libws2_32 and libiphlpapi under Windows, so this
hardcodes these libraries when building the ./configure test program.
These changes almost certainly break *nix, so should be fixed before merge.